refactor(compiler): move system include discovery into the compiler module
Discovery is a compiler query used by std registration regardless of
which command triggers it, so it belongs to the shared compiler tool.
